requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0

# Core dependencies (order matters)
accelerate>=0.20.0
transformers>=4.30.0
diffusers>=0.21.0
safetensors>=0.3.1

# LoRA specific
peft>=0.4.0

# Windows-specific (for CUDA support)
bitsandbytes-windows>=0.37.5

# Image processing
Pillow>=9.5.0

# Progress bars and utilities
tqdm>=4.65.0 